Always send notifications after commit if possible
This changes the behaviour of Devise::Async::Model to send
devise notifications after commits instead of right away when the
resource has not changed and `after_commit` is available.

The reason for that is Devise's behaviour when generating new
confirmation tokens: Devise generates a new token and calls `#save` on
the model. This marks the model as clean and `model.changed?` returns
false. This can happen inside a transaction.

The problem: if the Devise::Async worker is enqueued inside the
transaction, because `model#changed?` returns false, the transaction is
not committed and the changes are not visible to the worker yet.

This leads to a race condition where the worker might not see the
changes if it's fast enough.